backend/auth/google_auth.py
from fastapi import APIRouter, Request
from starlette.responses import RedirectResponse
from authlib.integrations.starlette_client import OAuth
from dotenv import load_dotenv
import os

load_dotenv()

oauth = OAuth()

oauth.register(
    name='google',
    client_id=os.getenv("GOOGLE_CLIENT_ID"),
    client_secret=os.getenv("GOOGLE_CLIENT_SECRET"),
    server_metadata_url='https://accounts.google.com/.well-known/openid-configuration',
    client_kwargs={
        'scope': 'openid email profile',
        'response_type': 'code'  # ✅ required to get id_token
    }
)

router = APIRouter()

@router.get("/auth/google")
async def login_via_google(request: Request):
    redirect_uri = os.getenv("GOOGLE_REDIRECT_URI")
    return await oauth.google.authorize_redirect(request, redirect_uri)

@router.get("/auth/google/callback")
async def google_callback(request: Request):
    token = await oauth.google.authorize_access_token(request)
    resp = await oauth.google.get("https://www.googleapis.com/oauth2/v3/userinfo", token=token)
  # ✅ manually fetch user info
    user = resp.json()
    return {
        "email": user.get("email"),
        "name": user.get("name"),
        "picture": user.get("picture")
    }